handle 0 vector for forces and more refactoring
